+++ spec/spec_helper.rb
@@ -46,21 +46,32 @@ def testing_ui
 end
 # rubocop:enable Lint/NestedMethodDefinition
 
-# Example environment (ENV) that would come from
-# running a PR on TravisCI
 def testing_env
   {
-    "HAS_JOSH_K_SEAL_OF_APPROVAL" => "true",
-    "TRAVIS_PULL_REQUEST" => "800",
-    "TRAVIS_REPO_SLUG" => "artsy/eigen",
-    "TRAVIS_COMMIT_RANGE" => "759adcbd0d8f...13c4dc8bb61d",
+    "BITRISE_PULL_REQUEST" => "4",
+    "BITRISE_IO" => "true",
+    "GIT_REPOSITORY_URL" => "git@github.com:artsy/eigen",
     "DANGER_GITHUB_API_TOKEN" => "123sbdq54erfsd3422gdfio"
+  }
+end
+
+def testing_env_for_gitlab
+  {
+    "BITRISE_PULL_REQUEST" => "4",
+    "BITRISE_IO" => "true",
+    "GIT_REPOSITORY_URL" => "git@gitlab.com:artsy/eigen",
+    "DANGER_GITLAB_API_TOKEN" => "123sbdq54erfsd3422gdfio"
   }
 end
 
 # A stubbed out Dangerfile for use in tests
 def testing_dangerfile
   env = Danger::EnvironmentManager.new(testing_env)
+  Danger::Dangerfile.new(env, testing_ui)
+end
+
+def testing_dangerfile_for_gitlab
+  env = Danger::EnvironmentManager.new(testing_env_for_gitlab)
   Danger::Dangerfile.new(env, testing_ui)
 end