+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Support GitLab and BitBucket server even if `inline_mode: false` is specified.
+
 ## 0.0.5
 
 ### Fixed